# docs/api/cotsel-dashboard-gateway.openapi.yml
openapi: 3.1.0
info:
  title: Cotsel Dashboard Gateway API
  version: 0.1.0
  summary: Contract-first admin/operator gateway for Cotsel-Dash.
  description: |
    Contract-first API specification for the Cotsel-Dash admin/operator dashboard to interact with Cotsel.

    Discovery outcome:
    - Dedicated in-repo dashboard gateway runtime: `gateway/` Express + TypeScript service.
    - Existing repo auth/session surface: `auth` service issues Bearer session tokens.
    - Existing repo governance source of truth: `contracts/src/AgroasysEscrow.sol`, `sdk/src/modules/adminSDK.ts`, and governance runbooks.
    - Existing repo compliance source of truth: policy/runbooks only; no dedicated compliance HTTP service exists yet.

    This specification is therefore contract-first for the dashboard gateway, but it is grounded in existing repo truth.

    Gateway authorization policy:
    - Auth uses existing auth-service Bearer sessions.
    - Gateway role mapping is derived from current auth role `admin`, which maps to `operator:read` and `operator:write`.
    - All mutation endpoints additionally require `GATEWAY_ENABLE_MUTATIONS=true` and membership in `GATEWAY_WRITE_ALLOWLIST`.
...
servers:
  - url: http://127.0.0.1:3600/api/dashboard-gateway/v1
    description: Local contract-first gateway base URL.
...
security:
  - sessionBearer: []
paths:
...
  /compliance/decisions:
    post:
      tags: [Compliance]
      summary: Create compliance decision record
      description: |
        Contract-first endpoint for compliance decisioning that gates oracle progression.

        This endpoint is grounded in `docs/runbooks/compliance-boundary-kyb-kyt-sanctions.md`, which defines fail-closed enforcement and audit requirements.

        The gateway owns the append-only compliance decision ledger in its Postgres database. Provider
        integrations remain out of scope; this endpoint records the authoritative control-plane decision
        used by the dashboard and gateway orchestration path. Off-chain attestation reference semantics
        used by later read surfaces are defined separately in `#/components/schemas/AttestationReference`.
      operationId: createComplianceDecision
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceDecisionCreateRequest'
      responses:
        '201':
          description: Decision recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceDecisionResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /compliance/decisions/{decisionId}:
    get:
      tags: [Compliance]
      summary: Get compliance decision by id
      operationId: getComplianceDecision
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - name: decisionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Compliance decision detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceDecisionResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /compliance/trades/{tradeId}:
    get:
      tags: [Compliance]
      summary: Get current compliance status for a trade
      operationId: getTradeComplianceStatus
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - $ref: '#/components/parameters/TradeId'
      responses:
        '200':
          description: Current compliance state for the trade.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceTradeStatusResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /compliance/trades/{tradeId}/attestation-status:
    get:
      tags: [Compliance]
      summary: Get read-only attestation status for a trade
      description: |
        Returns the latest attestation-backed compliance status for the trade.
        This is a read-only operator surface and does not revalidate the
        attestation against the issuer at request time. If the latest decision
        records provider outage or fail-closed denial without a fresh
        attestation payload, the response preserves the last known attestation
        metadata and marks the status as degraded or unavailable from the
        latest decision state.
      operationId: getTradeComplianceAttestationStatus
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - $ref: '#/components/parameters/TradeId'
      responses:
        '200':
          description: Current attestation-backed compliance status for the trade.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceTradeAttestationStatusResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /compliance/trades/{tradeId}/decisions:
    get:
      tags: [Compliance]
      summary: List compliance decisions for a trade
      operationId: listTradeComplianceDecisions
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - $ref: '#/components/parameters/TradeId'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Compliance decision history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceDecisionListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /compliance/trades/{tradeId}/block-oracle-progression:
    post:
      tags: [Compliance]
      summary: Block oracle progression for a trade
      description: |
        Contract-first off-chain control that prevents the oracle worker from progressing a trade.

        This does not change on-chain trade state directly. It is an operator/compliance control that must be consumed by the gateway/oracle orchestration path.

        The gateway stores oracle progression block state in its Postgres database as the canonical control-plane source of truth.
      operationId: blockOracleProgression
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/TradeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceOperationalControlRequest'
      responses:
        '202':
          description: Block accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceTradeStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /compliance/trades/{tradeId}/resume-oracle-progression:
    post:
      tags: [Compliance]
      summary: Resume oracle progression for a trade
      description: |
        Contract-first off-chain control that clears a prior oracle progression block.

        Resume requires an authorized gateway write principal together with the latest effective
        compliance decision evaluating to `ALLOW`. When the effective decision uses
        `reasonCode=CMP_OVERRIDE_ACTIVE`, the recorded override window must still be active.
      operationId: resumeOracleProgression
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XCorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/TradeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplianceOperationalControlRequest'
      responses:
        '202':
          description: Resume accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceTradeStatusResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
components:
  securitySchemes:
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: session-id
      description: |
        Session bearer token issued by the in-repo auth service.

        Production path:
        - trusted upstream session exchange via `/api/auth/v1/session/exchange/agroasys`

        Compatibility path:
        - `/api/auth/v1/login` after explicit wallet-signature login when that legacy flow is enabled

        Repo evidence:
        - `auth/src/api/controller.ts`
        - `auth/src/middleware/middleware.ts`
        - `auth/README.md`

        The gateway must resolve the session subject and enforce endpoint-level role policy. Governance and compliance mutations require operator/admin authorization.
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
      description: Client-generated request identifier for end-to-end traceability. Gateway should generate one if absent.
    XCorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
      description: Stable correlation identifier linking dashboard actions to downstream service logs, trade events, and incident artifacts.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
      description: Required for all mutation endpoints to deduplicate repeated operator intent.
...
    TradeId:
      name: tradeId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequestError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Authentication failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Authenticated subject lacks required role or scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Request conflicts with current state or prior idempotent request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error, timestamp]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: '#/components/schemas/ErrorObject'
        timestamp:
          type: string
          format: date-time
...
    EvidenceLink:
      type: object
      required: [kind, uri]
      properties:
        kind:
          type: string
          enum: [runbook, incident, ticket, tx, event, document, log, dashboard, other]
        uri:
          type: string
        note:
          type: string
    MutationAuditInput:
      type: object
      required: [reason, evidenceLinks, ticketRef]
      properties:
        reason:
          type: string
          minLength: 8
          maxLength: 2000
        evidenceLinks:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/EvidenceLink'
        ticketRef:
          type: string
          minLength: 2
          maxLength: 128
...
    AttestationStatus:
      type: string
      enum: [active, revoked, expired, unknown]
      description: |
        Canonical reference-state values for externally issued attestations.
        Verification freshness and outage handling are defined by the compliance
        runbooks and are intentionally separate from this schema-only status contract.
    AttestationIssuerRef:
      type: object
      required: [id, kind]
      properties:
        id:
          type: string
          description: Stable identifier for the issuer or integrating system.
        kind:
          type: string
          enum: [provider, service, operator, partner]
          description: Issuer class for the attestation boundary.
        displayName:
          type: string
          nullable: true
          description: Optional operator-facing label for the issuer.
    AttestationSubjectRef:
      type: object
      required: [type, reference]
      properties:
        type:
          type: string
          description: Scoped subject class such as `business`, `wallet`, `trade`, or `shipment`.
        reference:
          type: string
          description: Stable bounded subject reference. Must not contain raw PII or full provider dossiers.
    AttestationReference:
      type: object
      description: |
        Canonical off-chain attestation reference contract for operator surfaces.
        This schema intentionally carries only reference metadata and not the raw
        attestation payload. Verification freshness and outage interpretation are
        defined operationally in `docs/runbooks/compliance-boundary-kyb-kyt-sanctions.md`.
      required:
        - attestationId
        - attestationType
        - status
        - issuer
        - subjectRef
        - issuedAt
        - providerRef
        - evidenceRef
      properties:
        attestationId:
          type: string
          description: Stable identifier for the attestation in the issuing or integrating system.
        attestationType:
          type: string
          description: Bounded attestation category such as `kyb`, `kyt`, `sanctions_clearance`, or `inspection`.
        status:
          $ref: '#/components/schemas/AttestationStatus'
        issuer:
          $ref: '#/components/schemas/AttestationIssuerRef'
        subjectRef:
          $ref: '#/components/schemas/AttestationSubjectRef'
        issuedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        providerRef:
          type: string
          description: Provider or issuer lookup reference used for revalidation.
        evidenceRef:
          type: string
          description: Immutable evidence bundle or document reference linked to the attestation.
        referenceHash:
          type: string
          nullable: true
          description: Optional digest of the external attestation statement or evidence bundle.
    AttestationAvailability:
      type: string
      enum: [available, degraded, unavailable]
    AttestationFreshnessState:
      type: string
      enum: [current, stale, expired, unknown]
    ComplianceDecisionType:
      type: string
      enum: [KYB, KYT, SANCTIONS]
    ComplianceDecisionResult:
      type: string
      enum: [ALLOW, DENY]
      description: |
        Result states are intentionally limited to the policy-defined enforcement outcomes in
        `docs/runbooks/compliance-boundary-kyb-kyt-sanctions.md`.

        Emergency override is represented as `result=ALLOW` together with `reasonCode=CMP_OVERRIDE_ACTIVE`
        and the override audit fields below. Provider unavailability and unknown outcomes remain fail-closed
        as `result=DENY`.
    ComplianceBlockState:
      type: string
      enum: [not_blocked, blocked, resume_pending]
    ComplianceDecisionCreateRequest:
      type: object
      required:
        - tradeId
        - decisionType
        - result
        - reasonCode
        - provider
        - providerRef
        - subjectId
        - subjectType
        - correlationId
        - audit
      properties:
        tradeId:
          type: string
        decisionType:
          $ref: '#/components/schemas/ComplianceDecisionType'
        result:
          $ref: '#/components/schemas/ComplianceDecisionResult'
        reasonCode:
          type: string
          description: Stable code such as `CMP_PROVIDER_UNAVAILABLE` or `CMP_SANCTIONS_MATCH`.
        provider:
          type: string
        providerRef:
          type: string
        subjectId:
          type: string
        subjectType:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        overrideWindowEndsAt:
          type: string
          format: date-time
          nullable: true
          description: Used when `reasonCode=CMP_OVERRIDE_ACTIVE` to record the approved override window.
        correlationId:
          type: string
        attestation:
          $ref: '#/components/schemas/AttestationReference'
        audit:
          $ref: '#/components/schemas/MutationAuditInput'
    ComplianceDecision:
      type: object
      required:
        - decisionId
        - tradeId
        - decisionType
        - result
        - reasonCode
        - provider
        - providerRef
        - subjectId
        - subjectType
        - correlationId
        - decidedAt
        - audit
      properties:
        decisionId:
          type: string
        tradeId:
          type: string
        decisionType:
          $ref: '#/components/schemas/ComplianceDecisionType'
        result:
          $ref: '#/components/schemas/ComplianceDecisionResult'
        reasonCode:
          type: string
        provider:
          type: string
        providerRef:
          type: string
        subjectId:
          type: string
        subjectType:
          type: string
        riskLevel:
          type: string
          nullable: true
        correlationId:
          type: string
        decidedAt:
          type: string
          format: date-time
        overrideWindowEndsAt:
          type: string
          format: date-time
          nullable: true
        blockState:
          $ref: '#/components/schemas/ComplianceBlockState'
        attestation:
          $ref: '#/components/schemas/AttestationReference'
        audit:
          $ref: '#/components/schemas/MutationAuditRecord'
    ComplianceDecisionResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/ComplianceDecision'
        timestamp:
          type: string
          format: date-time
    ComplianceDecisionListResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ComplianceDecision'
            nextCursor:
              type: string
              nullable: true
        timestamp:
          type: string
          format: date-time
    ComplianceTradeStatus:
      type: object
      required:
        - tradeId
        - currentResult
        - oracleProgressionBlocked
        - latestDecisionId
        - updatedAt
      properties:
        tradeId:
          type: string
        currentResult:
          $ref: '#/components/schemas/ComplianceDecisionResult'
        oracleProgressionBlocked:
          type: boolean
        blockState:
          $ref: '#/components/schemas/ComplianceBlockState'
        latestDecisionId:
          type: string
        latestReasonCode:
          type: string
        latestProvider:
          type: string
        latestCorrelationId:
          type: string
        updatedAt:
          type: string
          format: date-time
    ComplianceTradeStatusResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/ComplianceTradeStatus'
        timestamp:
          type: string
          format: date-time
    ComplianceTradeAttestationStatus:
      type: object
      required:
        - tradeId
        - decisionId
        - decisionType
        - complianceResult
        - availability
        - freshness
        - verifiedAt
        - updatedAt
        - attestation
      properties:
        tradeId:
          type: string
        decisionId:
          type: string
        decisionType:
          $ref: '#/components/schemas/ComplianceDecisionType'
        complianceResult:
          $ref: '#/components/schemas/ComplianceDecisionResult'
        reasonCode:
          type: string
        availability:
          $ref: '#/components/schemas/AttestationAvailability'
        freshness:
          $ref: '#/components/schemas/AttestationFreshnessState'
        degradedReason:
          type: string
          nullable: true
        verifiedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        attestation:
          $ref: '#/components/schemas/AttestationReference'
    ComplianceTradeAttestationStatusResponse:
      type: object
      required: [success, data, timestamp]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/ComplianceTradeAttestationStatus'
        timestamp:
          type: string
          format: date-time
    ComplianceOperationalControlRequest:
      type: object
      required: [reasonCode, audit]
      properties:
        reasonCode:
          type: string
        audit:
          $ref: '#/components/schemas/MutationAuditInput'
        decisionId:
          type: string
          nullable: true
          description: Associated decision record when one exists.